🚚 Move tests to `src` directory.
