+++ package.json
@@ -5,7 +5,7 @@
   "scripts": {
     "build:formbuilder": "NODE_ENV=production rimraf build && webpack --config webpack.config.prod.js --optimize-minimize && cp public/index.prod.html build/index.html",
     "build:formbuilder-gh": "rimraf build && webpack --config webpack.config.github.js && cp public/index.prod.html build/index.html",
-    "lint": "eslint src test",
+    "lint": "eslint src src/test",
     "publish-to-gh-pages": "npm run build:formbuilder-gh && gh-pages --dist build/",
     "publish-to-npm": "npm run dist && npm publish",
     "start": "react-scripts start",
@@ -53,9 +53,6 @@
     "ts-jest": "^26.5.6",
     "typescript": "^4.2.4",
     "url-loader": "^4.1.1"
-  },
-  "directories": {
-    "test": "test"
   },
   "author": "Michał Bargiel <michal5bargiel@gmail.com>",
   "keywords": [
